+++ contracts/interfaces/IOracleRewardCB.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.28;
 
 interface IOracleRewardCB {
-  function newOracleRandomWords(uint randomWord) external;
+  function newOracleRandomWords(uint256 randomWord) external;
 }
